Consolidate tests, test TIFF as well
